refactor(test): audit attack.rs + adversarial.rs — merge, rehome misplaced tests
- Merge adversarial.rs into attack.rs (same theme: crafted proof rejection)
- Move 1 happy-path test to edge_cases.rs (root compression from TLA+)
- Move test_value_digest_accepts_wrong_path_as_exclusion to proof.rs
- Delete test_wrong_end_root_detected (duplicate of structural.rs)
- Delete test_root_shape_mismatch_low_range (redundant with root_structure_change)
- Rename db_a/db_b → source/target; add crafted prefix to tests 1-8
- Add test_out_of_range_root_expansion and test_root_value_change to edge_cases
- Delete adversarial.rs and remove mod adversarial from mod.rs
